refactor: Rename redux slice
