+++ src/store/index.ts
@@ -29,7 +29,7 @@ import { safeAppsSlice } from './safeAppsSlice'
 import { safeMessagesListener, safeMessagesSlice } from './safeMessagesSlice'
 import { pendingSafeMessagesSlice } from './pendingSafeMessagesSlice'
 import { batchSlice } from './batchSlice'
-import { undeployedSafeSlice } from '@/features/counterfactual/store/undeployedSafeSlice'
+import { undeployedSafesSlice } from '@/features/counterfactual/store/undeployedSafesSlice'
 
 const rootReducer = combineReducers({
   [chainsSlice.name]: chainsSlice.reducer,
@@ -50,7 +50,7 @@ const rootReducer = combineReducers({
   [safeMessagesSlice.name]: safeMessagesSlice.reducer,
   [pendingSafeMessagesSlice.name]: pendingSafeMessagesSlice.reducer,
   [batchSlice.name]: batchSlice.reducer,
-  [undeployedSafeSlice.name]: undeployedSafeSlice.reducer,
+  [undeployedSafesSlice.name]: undeployedSafesSlice.reducer,
 })
 
 const persistedSlices: (keyof PreloadedState<RootState>)[] = [
@@ -63,7 +63,7 @@ const persistedSlices: (keyof PreloadedState<RootState>)[] = [
   safeAppsSlice.name,
   pendingSafeMessagesSlice.name,
   batchSlice.name,
-  undeployedSafeSlice.name,
+  undeployedSafesSlice.name,
 ]
 
 export const getPersistedState = () => {
